+++ csfio.h
@@ -40,8 +40,8 @@ typedef struct {
 int csf_ctx_init(CSF_CTX **ctx_out, int *fh, unsigned char *keydata, int key_sz, int page_sz);
 int csf_truncate(CSF_CTX *ctx, int nByte);
 off_t csf_seek(CSF_CTX *ctx, off_t offset, int whence);
-int csf_read(CSF_CTX *ctx, void *buf, size_t nbyte);
-int csf_write(CSF_CTX *ctx, const void *buf, size_t nbyte);
+size_t csf_read(CSF_CTX *ctx, void *buf, size_t nbyte);
+size_t csf_write(CSF_CTX *ctx, const void *buf, size_t nbyte);
 int csf_ctx_destroy(CSF_CTX *ctx);
 
 #endif
